requirements.txt
appdirs==1.4.4
asgiref==3.3.4
backcall==0.2.0
beautifulsoup4==4.9.3
bs4==0.0.1
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
colorama==0.4.4
cryptography==3.4.6
decorator==4.4.2
defusedxml==0.7.1
distlib==0.3.1
dj-database-url==0.5.0
Django==3.2
django-allauth==0.43.0
django-braces==1.14.0
django-crispy-forms==1.11.1
django-heroku==0.3.1
filelock==3.0.12
gunicorn==20.1.0
html5lib==1.1
idna==2.10
ipykernel==5.4.3
ipython==7.19.0
ipython-genutils==0.2.0
jedi==0.18.0
jupyter-client==6.1.11
jupyter-core==4.7.0
oauthlib==3.1.0
parso==0.8.1
pickleshare==0.7.5
Pillow==8.1.2
prompt-toolkit==3.0.14
psycopg2==2.8.6
pycparser==2.20
Pygments==2.7.4
PyJWT==2.0.1

python-dateutil==2.8.1
python3-openid==3.2.0
pytz==2021.1

pyzmq==22.0.2
requests==2.25.1
requests-oauthlib==1.3.0
six==1.15.0
soupsieve==2.1
sqlparse==0.4.1
tornado==6.1
traitlets==5.0.5
urllib3==1.26.3
virtualenv==20.4.4
wcwidth==0.2.5
webencodings==0.5.1
whitenoise==5.2.0
